requirements.txt
customtkinter
requests
beautifulsoup4
google-genai
ollama
yfinance
python-docx
pandas-ta
cloudscraper
pyinstaller
matplotlib
PyPDF2
